Add tooltips for edit views of github widgets
Added tooltips that help with filling the configuration fields of the github widgets.
Therefore I created a edit view controller that fills the tooltip textes with html formatted strings.
The `tooltip-trigger="outsideClick" tooltip-is-open="tooltipIsOpen"  ng-mouseover="tooltipIsOpen = !tooltipIsOpen"` workaround is necessary because tooltips close if the mouse is not over the glyphicon and therefore any links would not be clickable.
Changing only the `tooltip-trigger` to `"outsideClick"`is not a suitable solution as the tooltip will only be displayed after the user clicks on the glyphicon.
Added a CHANGELOG.
Update dist folder.